Refactor import paths for admin actions
