+++ app/(admin)/inventory/page.tsx
