+++ app/(admin)/patients/page.tsx
