+++ app/auth/callback/route.ts
@@ -6,7 +6,7 @@ export async function GET(request: Request) {
   const { searchParams, origin } = new URL(request.url);
   const code = searchParams.get("code");
   // if "next" is in param, use it as the redirect URL
-  const next = searchParams.get("next") ?? "/admin";
+  const next = searchParams.get("next") ?? "/";
 
   if (code) {
     const supabase = createClient();
