+++ components/tables/patients-tables/column.tsx
@@ -3,7 +3,7 @@ import { ColumnDef } from "@tanstack/react-table";
 import { SquarePen } from "lucide-react";
 import { DeleteModal } from "@/components/modal/deleteModal";
 import { EditPatient } from "@/components/modal/patients/editPatient";
-import { deletePatient } from "@/app/admin/action";
+import { deletePatient } from "@/app/(admin)/action";
 import { mutate } from "swr";
 
 type Column = ColumnDef<PatientCol>;
